Switch services from H2 to MySQL database

+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `customer` (
-    `customer_id` INT AUTO_INCREMENT PRIMARY KEY,
+    `customer_id` BIGINT AUTO_INCREMENT PRIMARY KEY,
     `name` VARCHAR(100) NOT NULL,
     `email` VARCHAR(100) NOT NULL,
     `mobile_number` VARCHAR(20) NOT NULL,
@@ -10,8 +10,8 @@ CREATE TABLE IF NOT EXISTS `customer` (
 );
 
 CREATE TABLE IF NOT EXISTS `account` (
-    `customer_id` INT NOT NULL,
-    `account_number` INT AUTO_INCREMENT PRIMARY KEY,
+    `customer_id` BIGINT NOT NULL,
+    `account_number` BIGINT AUTO_INCREMENT PRIMARY KEY,
     `account_type` VARCHAR(100) NOT NULL,
     `branch_address` VARCHAR(200) NOT NULL,
     `created_at` DATE NOT NULL,
